refactor(china): extract the SSE/SZSE exchange transport into a shared module
The direct -> proxy -> edge ladder, bounded reads and failure classification
lived inside china-corporate-disclosures/adapters.mjs. #6155 adds a second
consumer on the same two hosts, so move the transport to
scripts/_china-exchange-transport.mjs verbatim and re-export the two names that
were already public. No behaviour change: tests/china-corporate-disclosures,
tests/source-proxy-fallback and the production-registration suite pass unchanged
(41/41).
